requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attributedict==0.3.0
attrs==23.2.0
auto-gptq @ https://github.com/jllllll/AutoGPTQ/releases/download/v0.6.0/auto_gptq-0.6.0+cu121-cp311-cp311-linux_x86_64.whl#sha256=80f44157c636a38ea12e0820ec681966310dfaa34b00724a176cd4c097b856d6
autoawq==0.1.8
bitsandbytes==0.42.0
blessings==1.7
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
contourpy==1.2.0
coverage==7.4.3
cramjam==2.8.1
ctransformers @ https://github.com/jllllll/ctransformers-cuBLAS-wheels/releases/download/AVX2/ctransformers-0.2.27+cu121-py3-none-any.whl#sha256=9be6bfa8ac9feb5b2d4c98fbf5ac90394bbfa5c406313f8161dca67b28333e51
cycler==0.12.1
DataProperty==1.0.1
datasets==2.17.1
deepdiff==6.7.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
exllamav2 @ https://github.com/oobabooga/exllamav2/releases/download/v0.0.14/exllamav2-0.0.14+cu121-cp311-cp311-linux_x86_64.whl#sha256=21a7ef889e3318968dac44b8018197ba5c59c674b5d21cf4147a553486d6e42e
fastapi==0.110.0
fastparquet==2024.2.0
ffmpy==0.3.2
filelock==3.13.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.3.4/flash_attn-2.3.4+cu122torch2.1cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=35f3335cc9b3c533e622fa5ea85908502f7aa558646523f6def10d1b53ca82e0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
gekko==1.0.6
gitdb==4.0.11
GitPython==3.1.42
gptq-for-llama @ https://github.com/jllllll/GPTQ-for-LLaMa-CUDA/releases/download/0.1.1/gptq_for_llama-0.1.1+cu121-cp311-cp311-linux_x86_64.whl#sha256=b6b0ce1b3b2568dff3c21d31956a82552e4eb6950c2a1f626767f9288ebc36d7
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.62.0
h11==0.14.0
hqq==0.1.3
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib_resources==6.1.2
inspecta==0.1.3
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llama_cpp_python @ https://github.com/oobabooga/llama-cpp-python-cuBLAS-wheels/releases/download/cpu/llama_cpp_python-0.2.50+cpuavx2-cp311-cp311-manylinux_2_31_x86_64.whl#sha256=cff3100779f71c8d2ce3ffbc350be63999749fe88dca4c0a10d6eee34672ea48
llama_cpp_python_cuda @ https://github.com/oobabooga/llama-cpp-python-cuBLAS-wheels/releases/download/textgen-webui/llama_cpp_python_cuda-0.2.50+cu121-cp311-cp311-manylinux_2_31_x86_64.whl#sha256=035584876fd88cd03dc2cedccc9933a2328cdffbb27833a7fe33aba68c7ba3b8
llama_cpp_python_cuda_tensorcores @ https://github.com/oobabooga/llama-cpp-python-cuBLAS-wheels/releases/download/textgen-webui/llama_cpp_python_cuda_tensorcores-0.2.50+cu121-cp311-cp311-manylinux_2_31_x86_64.whl#sha256=8908f740c8287d7c1155a98b73c5019469e7ae3bc772cf4bc8f27140575ad019
lm-eval==0.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mbstrdecoder==1.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
optimum==1.17.1
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathvalidate==3.2.0
peft==0.8.2
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pybind11==2.11.1
pycountry==23.12.11
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyproject-api==1.6.1
pytablewriter==1.2.0
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rootpath==0.1.1
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.18.0
sacrebleu==1.5.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.36.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.3.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tox==4.13.0
tqdm==4.66.2
tqdm-multiprocess==0.0.11
transformers==4.37.2
triton==2.2.0
typepy==1.3.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.25.1
wandb==0.16.3
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
